requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
Pillow>=10.0.0
nltk>=3.8.0
rouge-score>=0.1.2
datasets>=2.14.0
tensorboard>=2.14.0
tqdm>=4.65.0
